initial pytest support for surfstat
- move existsing SurfStatLinMod and SurfStatNorm tests to new
  directory tests
- adapt compatibility to pytest
- add pytest.ini

run like this:

 cd BrainStat/surfstat
 python3 -m pytest tests

The usage of pytest and the tests itself can be much more improved for
sure. This is just the initial step.
